Simplified gym wrapper, and added a second version of each environment which uses minimal_action_set.
