+++ gym_minatar/__init__.py
@@ -1,7 +1,11 @@
 from gym.envs.registration import register
 
 for game in ['Asterix', 'Breakout', 'Freeway', 'Seaquest', 'SpaceInvaders']:
-  register(
-    id='{}-MinAtar-v0'.format(game),
-    entry_point=f'gym_minatar.envs:{game}Env'
-)
+    register(
+        id='{}-MinAtar-v0'.format(game),
+        entry_point=f'gym_minatar.envs:{game}Env'
+    )
+    register(
+        id='{}-MinAtar-v1'.format(game),
+        entry_point=f'gym_minatar.envs:MinimalAction{game}Env'
+    )
