+++ gym_minatar/envs/__init__.py
@@ -1,6 +1,2 @@
 from gym_minatar.envs.base import BaseEnv
-from gym_minatar.envs.asterix import AsterixEnv
-from gym_minatar.envs.breakout import BreakoutEnv
-from gym_minatar.envs.freeway import FreewayEnv
-from gym_minatar.envs.seaquest import SeaquestEnv
-from gym_minatar.envs.space_invaders import SpaceInvadersEnv
+from gym_minatar.envs.game_envs import *
